Conflicts:
	pimcore/cli/classmap_generator.php
	pimcore/cli/send-newsletter.php
	pimcore/config/autoload-classmap.php
	pimcore/config/startup.php
	pimcore/config/texts/en.csv
	pimcore/lib/Pimcore.php
	pimcore/lib/Pimcore/API/Plugin/Broker.php
	pimcore/lib/Pimcore/Backup.php
	pimcore/lib/Pimcore/Controller/Action/Admin.php
	pimcore/lib/Pimcore/Controller/Action/Frontend.php
	pimcore/lib/Pimcore/Controller/Plugin/AdminButton.php
	pimcore/lib/Pimcore/Controller/Plugin/CssCombine.php
	pimcore/lib/Pimcore/Controller/Plugin/ErrorHandler.php
	pimcore/lib/Pimcore/Controller/Plugin/Frontend/Editmode.php
	pimcore/lib/Pimcore/Controller/Plugin/Targeting.php
	pimcore/lib/Pimcore/Controller/Plugin/Thumbnail.php
	pimcore/lib/Pimcore/Controller/Router/Route/Frontend.php
	pimcore/lib/Pimcore/Db/Profiler.php
	pimcore/lib/Pimcore/Document/Adapter/LibreOffice.php
	pimcore/lib/Pimcore/File.php
	pimcore/lib/Pimcore/File/Transfer/Adapter/Http.php
	pimcore/lib/Pimcore/Google/Api.php
	pimcore/lib/Pimcore/Helper/Dashboard.php
	pimcore/lib/Pimcore/Image/Adapter/Imagick.php
	pimcore/lib/Pimcore/Image/HtmlToImage.php
	pimcore/lib/Pimcore/Log/Simple.php
	pimcore/lib/Pimcore/Log/Writer/Mail.php
	pimcore/lib/Pimcore/Mail.php
	pimcore/lib/Pimcore/Model/Abstract.php
	pimcore/lib/Pimcore/Model/Cache.php
	pimcore/lib/Pimcore/Tool.php
	pimcore/lib/Pimcore/Tool/Admin.php
	pimcore/lib/Pimcore/Tool/Archive.php
	pimcore/lib/Pimcore/Tool/Console.php
	pimcore/lib/Pimcore/Tool/Less.php
	pimcore/lib/Pimcore/Tool/Mime.php
	pimcore/lib/Pimcore/Tool/Newsletter.php
	pimcore/lib/Pimcore/Tool/RestClient.php
	pimcore/lib/Pimcore/Tool/Session.php
	pimcore/lib/Pimcore/Update.php
	pimcore/lib/Pimcore/Version.php
	pimcore/lib/Pimcore/Video/Adapter/Ffmpeg.php
	pimcore/lib/Pimcore/View/Helper/PimcoreNavigation.php
	pimcore/lib/googleApiClient/cache/Google_ApcCache.php
	pimcore/lib/googleApiClient/contrib/Google_AdExchangeSellerService.php
	pimcore/lib/googleApiClient/contrib/Google_AnalyticsService.php
	pimcore/models/Asset.php
	pimcore/models/Asset/Document.php
	pimcore/models/Asset/Image.php
	pimcore/models/Asset/Image/Thumbnail/Config.php
	pimcore/models/Asset/Image/Thumbnail/Processor.php
	pimcore/models/Asset/Service.php
	pimcore/models/Asset/Video.php
	pimcore/models/Asset/Video/Thumbnail/Config.php
	pimcore/models/Asset/Video/Thumbnail/Processor.php
	pimcore/models/Asset/WebDAV/Service.php
	pimcore/models/Document.php
	pimcore/models/Document/Email/Log.php
	pimcore/models/Document/Tag/Href.php
	pimcore/models/Document/Tag/Multihref.php
	pimcore/models/Document/Tag/Renderlet.php
	pimcore/models/Document/Tag/Video.php
	pimcore/models/Element/Abstract.php
	pimcore/models/Element/Import/Service.php
	pimcore/models/Element/Recyclebin/Item.php
	pimcore/models/Element/Service.php
	pimcore/models/Object/Abstract.php
	pimcore/models/Object/Abstract/Resource.php
	pimcore/models/Object/Class.php
	pimcore/models/Object/Class/Data/Checkbox.php
	pimcore/models/Object/Class/Data/Date.php
	pimcore/models/Object/Class/Data/Datetime.php
	pimcore/models/Object/Class/Data/Fieldcollections.php
	pimcore/models/Object/Class/Data/Geobounds.php
	pimcore/models/Object/Class/Data/Geopoint.php
	pimcore/models/Object/Class/Data/Geopolygon.php
	pimcore/models/Object/Class/Data/Hotspotimage.php
	pimcore/models/Object/Class/Data/Href.php
	pimcore/models/Object/Class/Data/KeyValue.php
	pimcore/models/Object/Class/Data/Multihref.php
	pimcore/models/Object/Class/Data/Nonownerobjects.php
	pimcore/models/Object/Class/Data/Objectbricks.php
	pimcore/models/Object/Class/Data/Objects.php
	pimcore/models/Object/Class/Data/ObjectsMetadata.php
	pimcore/models/Object/Class/Data/Relations/Abstract.php
	pimcore/models/Object/Class/Data/StructuredTable.php
	pimcore/models/Object/Class/Data/Table.php
	pimcore/models/Object/Class/Resource.php
	pimcore/models/Object/Concrete.php
	pimcore/models/Object/Concrete/Resource.php
	pimcore/models/Object/Data/KeyValue.php
	pimcore/models/Object/Data/KeyValue/Entry.php
	pimcore/models/Object/Data/KeyValue/Resource.php
	pimcore/models/Object/Fieldcollection/Definition.php
	pimcore/models/Object/Folder.php
	pimcore/models/Object/KeyValue/KeyConfig.php
	pimcore/models/Object/Localizedfield.php
	pimcore/models/Object/Localizedfield/Resource.php
	pimcore/models/Object/Objectbrick/Data/Abstract.php
	pimcore/models/Object/Objectbrick/Data/Resource.php
	pimcore/models/Object/Objectbrick/Definition.php
	pimcore/models/Object/Service.php
	pimcore/models/Property.php
	pimcore/models/Schedule/Task/Executor.php
	pimcore/models/Site.php
	pimcore/models/Tool/CustomReport/Config.php
	pimcore/models/Tool/Newsletter/Config.php
	pimcore/models/Tool/Qrcode/Config.php
	pimcore/models/Tool/Setup.php
	pimcore/models/Tool/Setup/Resource.php
	pimcore/models/Tool/Tag/Config.php
	pimcore/models/Tool/Targeting/Persona.php
	pimcore/models/Translation/Abstract.php
	pimcore/models/Version.php
	pimcore/models/Webservice/Data/Object/Concrete.php
	pimcore/models/Webservice/Service.php
	pimcore/modules/admin/controllers/AdminButtonController.php
	pimcore/modules/admin/controllers/AssetController.php
	pimcore/modules/admin/controllers/DocumentController.php
	pimcore/modules/admin/controllers/ElementController.php
	pimcore/modules/admin/controllers/IndexController.php
	pimcore/modules/admin/controllers/LoginController.php
	pimcore/modules/admin/controllers/MiscController.php
	pimcore/modules/admin/controllers/ObjectController.php
	pimcore/modules/admin/controllers/ObjectHelperController.php
	pimcore/modules/admin/controllers/PageController.php
	pimcore/modules/admin/controllers/RecyclebinController.php
	pimcore/modules/admin/controllers/SettingsController.php
	pimcore/modules/admin/controllers/TranslationController.php
	pimcore/modules/admin/controllers/UserController.php
	pimcore/modules/admin/controllers/VariantsController.php
	pimcore/modules/admin/views/scripts/admin-button/promote.php
	pimcore/modules/admin/views/scripts/asset/show-version-image.php
	pimcore/modules/admin/views/scripts/index/index.php
	pimcore/modules/admin/views/scripts/object/diff-versions.php
	pimcore/modules/admin/views/scripts/object/preview-version.php
	pimcore/modules/extensionmanager/controllers/AdminController.php
	pimcore/modules/extensionmanager/controllers/DownloadController.php
	pimcore/modules/install/mysql/install.sql
	pimcore/modules/searchadmin/controllers/SearchController.php
	pimcore/modules/searchadmin/models/Search/Backend/Data.php
	pimcore/modules/webservice/controllers/RestController.php
	pimcore/modules/webservice/controllers/SoapController.php
	pimcore/static/js/frontend/targeting.js
	pimcore/static/js/pimcore/asset/tree.js
	pimcore/static/js/pimcore/document/tags/image.js
	pimcore/static/js/pimcore/document/tags/input.js
	pimcore/static/js/pimcore/document/tree.js
	pimcore/static/js/pimcore/object/classes/data/keyValue.js
	pimcore/static/js/pimcore/object/folder/search.js
	pimcore/static/js/pimcore/object/tags/hotspotimage.js
	pimcore/static/js/pimcore/object/tags/image.js
	pimcore/static/js/pimcore/object/tags/keyValue.js
	pimcore/static/js/pimcore/object/tree.js
	pimcore/static/js/pimcore/settings/system.js
	pimcore/static5/img/icon/page_white_code.png
	tests/SingleTest.php
	website_demo/build-dependency.txt
	website_demo/controllers/AdvancedController.php
	website_demo/controllers/NewsController.php
	website_demo/dump/create-dump.php
	website_demo/dump/data.sql
	website_demo/static/css/global.css
	website_demo/var/classes/Object/News.php
	website_demo/var/classes/definition_2.psf
	website_demo/views/layouts/layout.php
	website_demo/views/scripts/content/portal.php
	website_demo/views/scripts/includes/language.php
	website_demo/views/scripts/includes/sidebar.php
	website_example/views/scripts/default/default.php

+++ pimcore/lib/Pimcore/Cache/Backend/Memcached.php
@@ -9,15 +9,18 @@
  * It is also available through the world-wide-web at this URL:
  * http://www.pimcore.org/license
  *
- * @copyright  Copyright (c) 2009-2013 pimcore GmbH (http://www.pimcore.org)
+ * @copyright  Copyright (c) 2009-2014 pimcore GmbH (http://www.pimcore.org)
  * @license    http://www.pimcore.org/license     New BSD License
  */
 
+namespace Pimcore\Cache\Backend;
 
-class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
+use Pimcore\Resource;
+
+class Memcached extends \Zend_Cache_Backend_Memcached {
 
     /**
-     * @var Zend_Db_Adapter_Abstract
+     * @var \Zend_Db_Adapter_Abstract
      */
     protected $db;
 
@@ -25,6 +28,16 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
      * @var bool
      */
     protected $checkedCacheConsistency = false;
+
+    /**
+     * @param array $options
+     */
+    public function __construct(array $options = array()) {
+
+        $this->_options["tags_do_not_switch_to_innodb"] = null;
+
+        parent::__construct($options);
+    }
 
     /**
      * @return void
@@ -39,7 +52,7 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
 
             $res = $this->getDb()->fetchOne("SELECT id FROM cache_tags LIMIT 1");
             if(!$res) {
-                $this->clean(Zend_Cache::CLEANING_MODE_ALL);
+                $this->clean(\Zend_Cache::CLEANING_MODE_ALL);
             }
         }
     }
@@ -59,11 +72,13 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
     }
 
     /**
-     * @return Zend_Db_Adapter_Abstract
+     * @return \Zend_Db_Adapter_Abstract
      */
     protected function getDb () {
         if(!$this->db) {
-            $this->db = Pimcore_Resource::get();
+            // we're using a new mysql connection here to avoid problems with active (nested) transactions
+            \Logger::debug("Initialize dedicated MySQL connection for the cache adapter");
+            $this->db = Resource::getConnection();
         }
         return $this->db;
     }
@@ -85,14 +100,18 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
                         "tag" => $tag
                     ));
                 }
-                catch (Exception $e) {
+                catch (\Exception $e) {
                     if(strpos(strtolower($e->getMessage()), "is full") !== false) {
 
-                        Logger::warning($e);
+                        \Logger::warning($e);
 
-                        // it seems that the MEMORY table is on the limit an full
-                        // change the storage engine of the cache tags table to InnoDB
-                        $this->getDb()->query("ALTER TABLE `cache_tags` ENGINE=InnoDB");
+                        if($this->_options["tags_do_not_switch_to_innodb"]) {
+                            $this->clean();
+                        } else {
+                            // it seems that the MEMORY table is on the limit an full
+                            // change the storage engine of the cache tags table to InnoDB
+                            $this->getDb()->query("ALTER TABLE `cache_tags` ENGINE=InnoDB");
+                        }
 
                         // try it again
                         $tags[] = $tag;
@@ -104,8 +123,8 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
             }
             //$this->getDb()->commit();
 
-        } catch (Exception $e) {
-            Logger::error($e);
+        } catch (\Exception $e) {
+            \Logger::error($e);
         }
     }
 
@@ -146,19 +165,14 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
 
         $result = parent::save($data, $id, array(), $specificLifetime);
 
-        // hack may it works also without it
-        //$this->_memcache->delete($id);
-        // hack end
-
-        //$result = $this->_memcache->replace($id, array($data, time()), $flag, $lifetime);
-        //if( $result == false ) {
-        //    $result = $this->_memcache->set($id, array($data, time()), $flag, $lifetime);
-        //}
-        
-        
-        if (count($tags) > 0) {
-            $this->saveTags($id, $tags);
+        if($result) {
+            if (count($tags) > 0) {
+                $this->saveTags($id, $tags);
+            }
+        } else {
+            $this->remove($id);
         }
+
         return $result;
     }
 
@@ -170,14 +184,13 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
 
         $this->checkCacheConsistency();
 
+        $result = parent::remove($id);
 
         // using func_get_arg() to be compatible with the interface
         // when the 2ng argument is true, do not clean the cache tags
-        if(func_num_args() > 1 && func_get_arg(1) !== true) {
+        if($result && func_num_args() > 1 && func_get_arg(1) !== true) {
             $this->getDb()->delete("cache_tags", "id = '".$id."'");
         }
-
-        $result = parent::remove($id);
 
         // security check if the deletion fails
         if(!$result && $this->_memcache->get($id) !== false) {
@@ -202,34 +215,35 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
      * @param  array  $tags Array of tags
      * @return boolean True if no problem
      */
-    public function clean($mode = Zend_Cache::CLEANING_MODE_ALL, $tags = array()) {
+    public function clean($mode = \Zend_Cache::CLEANING_MODE_ALL, $tags = array()) {
 
         $this->checkCacheConsistency();
 
-        if ($mode == Zend_Cache::CLEANING_MODE_ALL) {
+        if ($mode == \Zend_Cache::CLEANING_MODE_ALL) {
             $this->clearTags();
             return $this->_memcache->flush();
         }
-        if ($mode == Zend_Cache::CLEANING_MODE_OLD) {
-            Logger::debug("Zend_Cache_Backend_Memcached::clean() : CLEANING_MODE_OLD is unsupported by the Memcached backend");
+        if ($mode == \Zend_Cache::CLEANING_MODE_OLD) {
+            \Logger::debug("Zend_Cache_Backend_Memcached::clean() : CLEANING_MODE_OLD is unsupported by the Memcached backend");
         }
-        if ($mode == Zend_Cache::CLEANING_MODE_MATCHING_TAG || $mode == Zend_Cache::CLEANING_MODE_MATCHING_ANY_TAG) {
+        if ($mode == \Zend_Cache::CLEANING_MODE_MATCHING_TAG || $mode == \Zend_Cache::CLEANING_MODE_MATCHING_ANY_TAG) {
             foreach ($tags as $tag) {
                 $items = $this->getItemsByTag($tag);
                 $quotedIds = array();
 
                 foreach ($items as $item) {
                     // We call delete directly here because the ID in the cache is already specific for this site
-                    $this->remove($item, true);
-                    $quotedIds[] = $this->getDb()->quote($item);
+                    if($this->remove($item, true)) {
+                        $quotedIds[] = $this->getDb()->quote($item);
+                    }
                 }
-                //$this->getDb()->delete("cache_tags", "tag = '".$tag."'");
+
                 if(count($quotedIds) > 0) {
                     $this->getDb()->delete("cache_tags", "id IN (" . implode(",", $quotedIds) . ")");
                 }
             }
         }
-        if ($mode == Zend_Cache::CLEANING_MODE_NOT_MATCHING_TAG) {
+        if ($mode == \Zend_Cache::CLEANING_MODE_NOT_MATCHING_TAG) {
             
             $condParts = array("1=1");
             foreach ($tags as $tag) {
@@ -249,7 +263,7 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
                 "id" => "___consistency_check___",
                 "tag" => "___consistency_check___"
             ));
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             // doesn't matter as long as the item exists
         }
